Bug #219869 fixed 1
