+++ .github/workflows/dev-build-deploy.yml
@@ -2,7 +2,7 @@ name: Deploy to DEV
 on:
   push:
     branches:
-      - release-2.0.6
+      - feat-sw-fix
 jobs:
   build:
     name: Generate Build and Deploy to DEV
