+++ apps/front-end/src/bootstrap.js
@@ -12,25 +12,21 @@ ReactDOM.render(
   document.getElementById("root")
 );
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://cra.link/PWA
-serviceWorkerRegistration.register();
+// Register service worker
+serviceWorkerRegistration.register({
+  onUpdate: registration => {
+    // Notify user about new updates
+    if (window.confirm("New version available! Would you like to update?")) {
+      if (registration && registration.waiting) {
+        registration.waiting.postMessage({ type: "SKIP_WAITING" });
+        registration.waiting.addEventListener("statechange", (e) => {
+          if (e.target.state === "activated") {
+            window.location.reload();
+          }
+        });
+      }
+    }
+  }
+});
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-
-//cache clear 1.5.13
-
-/*if ("serviceWorker" in navigator) {
-  navigator.serviceWorker
-    .register("/service-worker.js")
-    .then((registration) => {
-      console.log("Service Worker registered with scope:", registration.scope);
-    })
-    .catch((error) => {
-      console.error("Service Worker registration failed:", error);
-    });
-}*/
