+++ apps/front-end/src/pages/front-end/Dashboard.js
@@ -611,7 +611,7 @@ export default function Dashboard({ userTokenInfo, footerLinks }) {
                 resizeMode="contain"
               />
               <FrontEndTypo.H1 color="textMaroonColor.400" pl="1">
-                {t("WELCOME")} {facilitator?.first_name},
+                {t("WELCOME")} {facilitator?.first_name}, v1 27 may 24
               </FrontEndTypo.H1>
             </HStack>
             {events?.length ? (
