+++ apps/front-end/src/service-worker.js
@@ -1,25 +1,15 @@
 /* eslint-disable no-restricted-globals */
-
-// This service worker can be customized!
-// See https://developers.google.com/web/tools/workbox/modules
-// for the list of available Workbox modules, or add any other
-// code you'd like.
-// You can also remove this file if you'd prefer not to use a
-// service worker, and the Workbox build step will be skipped.
-
 import { clientsClaim } from "workbox-core";
 import { ExpirationPlugin } from "workbox-expiration";
 import { precacheAndRoute, createHandlerBoundToURL } from "workbox-precaching";
 import { registerRoute } from "workbox-routing";
-import { StaleWhileRevalidate } from "workbox-strategies";
+import { StaleWhileRevalidate, NetworkFirst } from "workbox-strategies";
 
-// public/service-worker.js
+// Import Workbox from CDN
+importScripts("https://storage.googleapis.com/workbox-cdn/releases/6.1.5/workbox-sw.js");
 
-importScripts(
-  "https://storage.googleapis.com/workbox-cdn/releases/4.3.1/workbox-sw.js"
-);
-
-importScripts("/precache-manifest.a4724df64b745797c25a9173550ba2d3.js");
+// Claim clients immediately
+clientsClaim();
 
 self.addEventListener("message", (event) => {
   if (event.data && event.data.type === "SKIP_WAITING") {
@@ -27,190 +17,38 @@ self.addEventListener("message", (event) => {
   }
 });
 
-workbox.core.clientsClaim();
-
-/**
- * The workboxSW.precacheAndRoute() method efficiently caches and responds to
- * requests for URLs in the manifest.
- */
-self.__precacheManifest = [].concat(self.__precacheManifest || []);
-workbox.precaching.precacheAndRoute(self.__precacheManifest, {});
-
-workbox.routing.registerNavigationRoute(
-  workbox.precaching.getCacheKeyForURL("/index.html"),
-  {
-    blacklist: [/^\/_/, /\/[^/?]+\.[^/]+$/],
-  }
-);
-
-// Precache all of the assets generated by your build process.
-// Their URLs are injected into the manifest variable below.
-// This variable must be present somewhere in your service worker file,
-// even if you decide not to use precaching. See https://cra.link/PWA
+// Precache files
 precacheAndRoute(self.__WB_MANIFEST);
 
-// Set up App Shell-style routing, so that all navigation requests
-// are fulfilled with your index.html shell. Learn more at
-// https://developers.google.com/web/fundamentals/architecture/app-shell
-const fileExtensionRegexp = new RegExp("/[^/?]+\\.[^/]+$");
+// Use NetworkFirst strategy for HTML files to ensure fresh content
 registerRoute(
-  // Return false to exempt requests from being fulfilled by index.html.
-  ({ request, url }) => {
-    // If this isn't a navigation, skip.
-    if (request.mode !== "navigate") {
-      return false;
-    } // If this is a URL that starts with /_, skip.
-
-    if (url.pathname.startsWith("/_")) {
-      return false;
-    } // If this looks like a URL for a resource, because it contains // a file extension, skip.
-
-    // If this looks like a URL for a resource and doesn't end with .html, skip.
-    // if (
-    //   url.pathname.match(fileExtensionRegexp) &&
-    //   !url.pathname.endsWith(".html")
-    // ) {
-    //   return false;
-    // }
-    // Return true to signal that we want to use the handler.
-
-    return true;
-  },
-  createHandlerBoundToURL(process.env.PUBLIC_URL + "/index.html")
+  ({ request }) => request.mode === "navigate",
+  new NetworkFirst({
+    cacheName: "pages-cache",
+    plugins: [
+      new ExpirationPlugin({ maxEntries: 50 })
+    ],
+  })
 );
 
-// An example runtime caching route for requests that aren't handled by the
-// precache, in this case same-origin .png requests like those from in public/
+// Use StaleWhileRevalidate for images and other assets
 registerRoute(
-  // Add in any other file extensions or routing criteria as needed.
-  ({ url }) =>
-    url.origin === self.location.origin && url.pathname.endsWith(".png"), // Customize this strategy as needed, e.g., by changing to CacheFirst.
+  ({ request }) => request.destination === "image",
   new StaleWhileRevalidate({
-    cacheName: "images",
+    cacheName: "images-cache",
     plugins: [
-      // Ensure that once this runtime cache reaches a maximum size the
-      // least-recently used images are removed.
       new ExpirationPlugin({ maxEntries: 50 }),
     ],
   })
 );
 
+// Default fallback for other requests
 registerRoute(
-  // Return false to exempt requests from being fulfilled by index.html.
-  ({ request, url }) => {
-    // If this isn't a navigation, skip.
-    if (request.mode !== "navigate") {
-      return false;
-    }
-
-    // If this is a URL that starts with /_, skip.
-    if (url.pathname.startsWith("/_")) {
-      return false;
-    }
-
-    // If this looks like a URL for a resource and doesn't end with .html, skip.
-    // If this looks like a URL for a resource and doesn't end with .html, skip.
-    // if (
-    //   url.pathname.match(fileExtensionRegexp) &&
-    //   !url.pathname.endsWith(".html")
-    // ) {
-    //   return false;
-    // }
-    // Return true to signal that we want to use the handler.
-    return true;
-  },
-  createHandlerBoundToURL(process.env.PUBLIC_URL + "/index.html")
+  ({ request }) => request.destination !== "image" && request.mode !== "navigate",
+  new StaleWhileRevalidate({
+    cacheName: "default-cache",
+    plugins: [
+      new ExpirationPlugin({ maxEntries: 50 }),
+    ],
+  })
 );
-
-// This allows the web app to trigger skipWaiting via
-// registration.waiting.postMessage({type: 'SKIP_WAITING'})
-self.addEventListener("message", (event) => {
-  if (event.data && event.data.type === "SKIP_WAITING") {
-    self.skipWaiting();
-  }
-});
-
-// Any other custom service worker logic can go here.
-
-//old code
-/*self.addEventListener("fetch", (event) => {
-  if (event.request.mode === "navigate") {
-    // If it's a navigation request (HTML file), don't cache it
-    return;
-  }
-});*/
-
-//new code
-/*const CACHE_NAME = "eg-pragati-v1";
-const urlsToCache = [
-  "/",
-  "/index.html",
-  "/offline.html", // Add your custom offline page
-  // Add other static assets to cache
-];
-
-self.addEventListener("install", (event) => {
-  event.waitUntil(
-    caches
-      .open(CACHE_NAME)
-      .then((cache) => cache.addAll(urlsToCache))
-      .then(() => self.skipWaiting())
-  );
-});
-
-self.addEventListener("activate", (event) => {
-  event.waitUntil(
-    caches.keys().then((cacheNames) => {
-      return Promise.all(
-        cacheNames.map((cacheName) => {
-          if (cacheName !== CACHE_NAME) {
-            return caches.delete(cacheName);
-          }
-        })
-      );
-    })
-  );
-  self.clients.claim();
-});
-
-self.addEventListener("fetch", (event) => {
-  event.respondWith(
-    caches
-      .match(event.request)
-      .then((response) => {
-        // If resource is cached, return it
-        if (response) {
-          return response;
-        }
-        // If offline and request for HTML, return cached index.html
-        if (!navigator.onLine && event.request.url.endsWith("/index.html")) {
-          return caches.match("/index.html");
-        }
-        // If offline and request for other resources, return custom offline response
-        if (!navigator.onLine) {
-          return new Response("Offline", {
-            status: 503,
-            statusText: "Service Unavailable",
-          });
-        }
-        // If online and resource not found in cache, fetch from network
-        return fetch(event.request).then((networkResponse) => {
-          const clonedResponse = networkResponse.clone();
-          caches.open(CACHE_NAME).then((cache) => {
-            cache.put(event.request, clonedResponse);
-          });
-          return networkResponse;
-        });
-      })
-      .catch(() => {
-        // If fetch fails and resource not in cache, return custom offline response
-        if (event.request.url.endsWith("/index.html")) {
-          return caches.match("/index.html");
-        }
-        return new Response("Offline", {
-          status: 503,
-          statusText: "Service Unavailable",
-        });
-      })
-  );
-});*/
